move functions to utils/transaction
